Fixing rate types for log_rates()
git-svn-id: https://lwes.svn.sourceforge.net/svnroot/lwes/lwes-journaller/trunk@534 a2f82657-cdd2-4550-bd36-68a8e7111808

+++ src/stats.c
@@ -142,7 +142,7 @@ void dequeuer_stats_record_loss (struct dequeuer_stats* st)
 }
 
 
-static void log_rates(log_mask_t level, const char* file, int line, long long bps, long long pps,
+static void log_rates(log_mask_t level, const char* file, int line, double bps, double pps,
                       const char* notes)
 {
   if ( bps > 1000000. )
